+++ tests/integration/testapp/app.py
@@ -146,8 +146,12 @@ def multifile():
 
 @app.route('/custom-response', methods=['GET'])
 def custom_response():
-    return Response(status_code=204, body='',
-                    headers={'Content-Type': 'text/plain'})
+    return Response(
+        status_code=204,
+        body='',
+        headers={'Content-Type': 'text/plain'},
+        multi_value_headers={'Set-Cookie': ['key=value', 'foo=bar']}
+    )
 
 
 @app.route('/api-key-required', methods=['GET'], api_key_required=True)